Silence build-time warning noise from vendor SDK code and callback-style interfaces
Discovered on Windows during 0.4.0 staging verification: with Arduino
IDE's "Compiler warnings" preference set above "None" (its default is
"None", so macOS/Linux testers hadn't seen this), -Wall/-Wextra
surfaces ~49 warnings, all originating from either the NXP MCUXpresso
SDK's own vendor driver files (unused-parameter, missing-field-
initializers, old-style-declaration -- generic driver APIs with
per-variant unused arguments) or this project's own callback/virtual-
override-style methods with the same shape (BusIn/BusOut::config(),
I3C's SDK-mandated callback signatures, I2C's CCC callback stub).

Fixed:
- platform.txt: added -Wno-unused-parameter/-Wno-missing-field-
  initializers/-Wno-old-style-declaration to compiler.c.flags (vendor
  SDK is all .c), and -Wno-unused-parameter/-Wno-missing-field-
  initializers to compiler.cpp.flags (this project's own r01lib code).
  Matches the existing -Wno-address-of-packed-member precedent already
  used for SD library compatibility -- vendor files stay byte-for-byte
  unmodified, and only these specific low-value warning classes are
  suppressed, not warnings generally.
- i3c.cpp: the only warnings NOT from this class -- three
  enumerated/non-enumerated ternary warnings in I3C::I3C() mixing
  uint32_t parameters with enum FREQ constants -- fixed with explicit
  (uint32_t) casts (behavior-preserving; unscoped enums already
  convert implicitly, this just makes the ternary's result type
  unambiguous).

Verified: zero warnings on both boards with `arduino-cli compile
--warnings all` from a fully cleared build cache (the strictest
level, stricter than any real user is likely to select). Binary size
unchanged (+32/+28 bytes, from the casts only -- the -Wno- flags don't
affect codegen). Full regression sweep across all 61 example sketches
on both boards: no new failures.

+++ hardware/nxp/mcx/cores/arduino/r01lib/i3c.cpp
@@ -79,9 +79,9 @@ I3C::I3C( int sda, int scl, uint32_t i2c_freq, uint32_t i3c_od_freq, uint32_t i3
 	
 	I3C_MasterGetDefaultConfig( &masterConfig );
 
-	masterConfig.baudRate_Hz.i2cBaud          = i2c_freq    ? i2c_freq    : I2C::FREQ;
-	masterConfig.baudRate_Hz.i3cOpenDrainBaud = i3c_od_freq ? i3c_od_freq : OD_FREQ;
-	masterConfig.baudRate_Hz.i3cPushPullBaud  = i3c_pp_freq ? i3c_pp_freq : PP_FREQ;
+	masterConfig.baudRate_Hz.i2cBaud          = i2c_freq    ? i2c_freq    : (uint32_t)I2C::FREQ;
+	masterConfig.baudRate_Hz.i3cOpenDrainBaud = i3c_od_freq ? i3c_od_freq : (uint32_t)OD_FREQ;
+	masterConfig.baudRate_Hz.i3cPushPullBaud  = i3c_pp_freq ? i3c_pp_freq : (uint32_t)PP_FREQ;
 	masterConfig.enableOpenDrainStop          = false;
 	masterConfig.disableTimeout               = true;
